requirements.txt
opencv-python==4.7.0.72
ultralytics==8.2.93
numpy==1.24.4
openai==1.45.0
timm==1.0.9
python-dotenv==1.0.1
pyaudio==0.2.14
speechrecognition==3.10.4
faster_whisper==1.0.3
pillow==10.4.0
